# docs/contracts/maekon-web.v1.openapi.yaml
openapi: 3.0.3
info:
  title: Maekon Local Web API
  version: "v1"
  description: |
    Auto-generated from docs/contracts/http-interface-manifest.v1.json.
    Request/response bodies are real schemars-generated DTO schemas
    (full disclosure; the API is loopback-only + token-gated). Cross-crate
    maekon-core field types are rendered as opaque objects (build containment).
  x-maekon-document-version: 10
  x-maekon-updated-at: "2026-07-23"
  x-maekon-routes-file: "crates/maekon-web/src/routes.rs"
  x-maekon-contracts-crate: "crates/maekon-api-contracts"
servers:
  - url: /
paths:
...
  "/api/backup":
    get:
      tags:
        - backup
      operationId: backup_get_api_backup
      summary: "GET /api/backup"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/dashboard/day":
    get:
      tags:
        - dashboard
      operationId: dashboard_get_api_dashboard_day
      summary: "GET /api/dashboard/day"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/events":
    get:
      tags:
        - events
      operationId: events_get_api_events
      summary: "GET /api/events"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/export/events":
    get:
      tags:
        - export
      operationId: export_get_api_export_events
      summary: "GET /api/export/events"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/export/frames":
    get:
      tags:
        - export
      operationId: export_get_api_export_frames
      summary: "GET /api/export/frames"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/export/full":
    get:
      tags:
        - export
      operationId: export_get_api_export_full
      summary: "GET /api/export/full"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/frames":
    get:
      tags:
        - frames
      operationId: frames_get_api_frames
      summary: "GET /api/frames"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/frames/{id}/image":
    get:
      tags:
        - frames
      operationId: frames_get_api_frames_id_image
      summary: "GET /api/frames/{id}/image"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/search":
    get:
      tags:
        - search
      operationId: search_get_api_search
      summary: "GET /api/search"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/semantic-search":
    get:
      tags:
        - semantic_search
      operationId: semantic_search_get_api_semantic_search
      summary: "GET /api/semantic-search"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SemanticSearchResult'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/semantic-search/capabilities":
    get:
      tags:
        - semantic_search
      operationId: semantic_search_get_api_semantic_search_capabilities
      summary: "GET /api/semantic-search/capabilities"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticSearchCapabilities'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/settings":
    get:
      tags:
        - settings
      operationId: settings_get_api_settings
      summary: "GET /api/settings"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - settings
      operationId: settings_post_api_settings
      summary: "POST /api/settings"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppSettings'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/stats/summary":
    get:
      tags:
        - stats
      operationId: stats_get_api_stats_summary
      summary: "GET /api/stats/summary"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailySummaryResponse'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/timeline":
    get:
      tags:
        - timeline
      operationId: timeline_get_api_timeline
      summary: "GET /api/timeline"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
...
    AiProviderSettings: {"properties":{"access_mode":{"type":"string"},"active_profile_id":{"type":["string","null"]},"bypass_pii_filter_for_external_ocr":{"default":false,"type":"boolean"},"external_data_policy":{"type":"string"},"fallback_to_local":{"type":"boolean"},"llm_api":{"anyOf":[{"$ref":"#/components/schemas/ExternalApiSettings"},{"type":"null"}]},"llm_provider":{"type":"string"},"ocr_api":{"anyOf":[{"$ref":"#/components/schemas/ExternalApiSettings"},{"type":"null"}]},"ocr_provider":{"type":"string"},"ocr_validation":{"$ref":"#/components/schemas/OcrValidationSettings","default":{"enabled":true,"max_invalid_ratio":0.6,"min_confidence":0.25}},"saved_profiles":{"default":[],"items":{"$ref":"#/components/schemas/SavedAiProviderProfile"},"type":"array"},"scene_action_override":{"$ref":"#/components/schemas/SceneActionOverrideSettings","default":{"approved_by":"","enabled":false,"reason":""}},"scene_intelligence":{"$ref":"#/components/schemas/SceneIntelligenceSettings","default":{"allow_action_execution":false,"calibration_enabled":true,"calibration_min_avg_confidence":0.55,"calibration_min_elements":8,"enabled":true,"max_elements":120,"min_confidence":0.35,"overlay_enabled":true}}},"required":["access_mode","ocr_provider","llm_provider","external_data_policy","fallback_to_local"],"type":"object"}
...
    AiSessionSettings: {"properties":{"health_check_interval_secs":{"format":"uint64","minimum":0,"type":"integer"},"idle_timeout_secs":{"format":"uint64","minimum":0,"type":"integer"},"max_concurrent_sessions":{"format":"uint32","minimum":0,"type":"integer"},"max_history_turns":{"format":"uint32","minimum":0,"type":"integer"},"max_output_tokens":{"default":4096,"format":"uint32","minimum":0,"type":"integer"},"max_retries":{"format":"uint32","minimum":0,"type":"integer"},"session_timeout_secs":{"format":"uint64","minimum":0,"type":"integer"},"thinking":true},"required":["max_concurrent_sessions","idle_timeout_secs","session_timeout_secs","max_retries","max_history_turns","health_check_interval_secs"],"type":"object"}
    AnalysisSettings: {"properties":{"auto_tuner_enabled":{"description":"Whether the EMA-based auto-tuner (drift detection + re-clustering) is active.","type":"boolean"},"embedding_enabled":{"type":"boolean"},"enabled":{"type":"boolean"},"gui_intelligence_enabled":{"type":"boolean"},"interval_secs":{"format":"uint64","minimum":0,"type":"integer"},"llm_summary_enabled":{"description":"Whether the local LLM generates a natural-language daily-digest narrative\nbefore embedding (maps to `analysis.embedding.llm_summary_enabled`). The\nnarrative pipeline only runs when analysis, embedding, and this flag are\nall enabled; the rule-based digest always runs regardless.","type":"boolean"},"max_suggestions":{"format":"uint32","minimum":0,"type":"integer"},"min_confidence":{"format":"double","type":"number"},"regime_detection_interval_hours":{"default":2,"description":"Regime-detection interval (`analysis.tiered_memory.regime_detection_interval_hours`).\n\nThe AdvancedTab control for this value previously (#9629) read a nested\nkey the server never sent and wrote a flat key serde dropped — a full\nround-trip break. A flat contract field fixes both directions.","format":"int64","type":"integer"},"text_intelligence_enabled":{"type":"boolean"},"tiered_memory_enabled":{"default":false,"description":"Master switch for the tiered-memory pipeline (`analysis.tiered_memory.enabled`).\n\nPowers /day, /week, recalibration segments, coaching goal progress, and\nhabit streaks. Before this field existed (#9629) the switch was readable\nat boot but writable NOWHERE (no REST, no IPC, no UI), leaving those\nsurfaces permanently empty. `serde(default)` keeps payloads from older\nfrontends valid.","type":"boolean"}},"required":["enabled","interval_secs","min_confidence","max_suggestions","embedding_enabled","llm_summary_enabled","gui_intelligence_enabled","text_intelligence_enabled","auto_tuner_enabled"],"type":"object"}
...
    AppSettings: {"properties":{"ai_provider":{"$ref":"#/components/schemas/AiProviderSettings","default":{"access_mode":"provider_api_key","bypass_pii_filter_for_external_ocr":false,"external_data_policy":"PiiFilterStrict","fallback_to_local":true,"llm_provider":"Local","ocr_provider":"Local","ocr_validation":{"enabled":true,"max_invalid_ratio":0.6,"min_confidence":0.25},"saved_profiles":[],"scene_action_override":{"approved_by":"","enabled":false,"reason":""},"scene_intelligence":{"allow_action_execution":false,"calibration_enabled":true,"calibration_min_avg_confidence":0.55,"calibration_min_elements":8,"enabled":true,"max_elements":120,"min_confidence":0.35,"overlay_enabled":true}}},"ai_session":{"$ref":"#/components/schemas/AiSessionSettings","default":{"health_check_interval_secs":30,"idle_timeout_secs":300,"max_concurrent_sessions":3,"max_history_turns":100,"max_output_tokens":4096,"max_retries":3,"session_timeout_secs":600}},"allow_external":{"type":"boolean"},"analysis":{"$ref":"#/components/schemas/AnalysisSettings","default":{"auto_tuner_enabled":true,"embedding_enabled":true,"enabled":true,"gui_intelligence_enabled":true,"interval_secs":60,"llm_summary_enabled":false,"max_suggestions":5,"min_confidence":0.5,"regime_detection_interval_hours":2,"text_intelligence_enabled":true,"tiered_memory_enabled":false}},"audio":{"$ref":"#/components/schemas/AudioSettings","default":{"cloud_api_key":"","cloud_stt_endpoint":"https://api.openai.com/v1/audio/transcriptions","cloud_timeout_secs":10,"enabled":false,"language":"auto","max_recording_secs":60,"mic_input_mode":"push_to_talk","model_size":"base","stt_provider":"local","vad_min_speech_ms":300,"vad_silence_ms":800,"vad_threshold":0.019999999552965164,"whisper_model_path":""}},"automation":{"$ref":"#/components/schemas/AutomationSettings","default":{"enabled":false}},"capture_enabled":{"type":"boolean"},"coaching":{"$ref":"#/components/schemas/CoachingSettings","default":{"enabled":false,"locale":"en","overlay_mode":"Minimal","profiles":{"ContextRestore":{"enabled":true,"min_interval_secs":300},"DeepWorkCoach":{"enabled":true,"min_interval_secs":300},"FocusGuard":{"enabled":true,"min_interval_secs":300},"GoalTracker":{"enabled":true,"min_interval_secs":300},"TimeAware":{"enabled":true,"min_interval_secs":300}},"quiet_hours":[],"regime_goals":{},"tone":"Gentle"}},"focus_auto":{"$ref":"#/components/schemas/FocusAutoSettings","default":{"cooldown_secs":300,"duration_minutes":25,"enabled":false,"trigger_apps":[],"trigger_schedules":[]}},"idle_threshold_secs":{"format":"uint32","minimum":0,"type":"integer"},"indicator":{"$ref":"#/components/schemas/IndicatorSettings","default":{"border_opacity":0.6000000238418579,"show_border":true,"show_panel":true}},"integration":{"$ref":"#/components/schemas/IntegrationSettings","default":{"auth_profile_kind":"none","enabled":false,"request_timeout_secs":30,"sync_interval_secs":60}},"max_storage_mb":{"format":"uint32","minimum":0,"type":"integer"},"metrics_interval_secs":{"format":"uint32","minimum":0,"type":"integer"},"monitor":{"$ref":"#/components/schemas/MonitorControlSettings","default":{"input_activity":true,"privacy_mode":false,"process_monitoring":true}},"network":{"$ref":"#/components/schemas/NetworkSettings","default":{"grpc_enabled":false,"grpc_endpoint":"http://localhost:50051","request_timeout_ms":30000,"server_base_url":"http://localhost:8000","tls_enabled":false}},"notification":{"$ref":"#/components/schemas/NotificationSettings","default":{"enabled":false,"high_usage_notification":false,"high_usage_threshold":0,"idle_notification":false,"idle_notification_mins":0,"long_session_mins":0,"long_session_notification":false}},"privacy":{"$ref":"#/components/schemas/PrivacySettings","default":{"auto_exclude_sensitive":false,"excluded_app_patterns":[],"excluded_apps":[],"excluded_title_patterns":[],"pii_filter_level":""}},"process_interval_secs":{"format":"uint32","minimum":0,"type":"integer"},"retention_days":{"format":"uint32","minimum":0,"type":"integer"},"sandbox":{"$ref":"#/components/schemas/SandboxSettings","default":{"allow_network":false,"allowed_read_paths":[],"allowed_write_paths":[],"enabled":false,"max_cpu_time_ms":0,"max_memory_bytes":0,"profile":"Standard"}},"schedule":{"$ref":"#/components/schemas/ScheduleSettings","default":{"active_days":["Mon","Tue","Wed","Thu","Fri"],"active_end_hour":18,"active_hours_enabled":false,"active_start_hour":9,"pause_on_battery_saver":false,"pause_on_screen_lock":true}},"suggestion":{"$ref":"#/components/schemas/SuggestionSettings","default":{"enabled":false}},"sync":{"$ref":"#/components/schemas/SyncSettings","default":{"compression_enabled":true,"device_name":"","enabled":false,"interval_secs":300,"lan_advertise":false,"transport":"none"}},"telemetry":{"$ref":"#/components/schemas/TelemetrySettings","default":{"crash_reports":false,"enabled":false,"performance_metrics":false,"usage_analytics":false}},"update":{"$ref":"#/components/schemas/UpdateSettings","default":{"auto_install":false,"channel":"stable","check_interval_hours":24,"enabled":true}},"web_port":{"format":"uint16","maximum":65535,"minimum":0,"type":"integer"}},"required":["retention_days","max_storage_mb","web_port","allow_external","capture_enabled","idle_threshold_secs","metrics_interval_secs","process_interval_secs"],"type":"object"}
...
    AudioSettings: {"properties":{"cloud_api_key":{"default":"","type":"string"},"cloud_stt_endpoint":{"default":"https://api.openai.com/v1/audio/transcriptions","type":"string"},"cloud_timeout_secs":{"default":10,"format":"uint32","minimum":0,"type":"integer"},"enabled":{"default":false,"type":"boolean"},"language":{"default":"auto","type":"string"},"max_recording_secs":{"default":60,"format":"uint32","minimum":0,"type":"integer"},"mic_input_mode":{"default":"push_to_talk","type":"string"},"model_size":{"default":"base","type":"string"},"stt_provider":{"default":"local","type":"string"},"vad_min_speech_ms":{"default":300,"format":"uint32","minimum":0,"type":"integer"},"vad_silence_ms":{"default":800,"format":"uint32","minimum":0,"type":"integer"},"vad_threshold":{"default":0.019999999552965164,"format":"float","type":"number"},"whisper_model_path":{"default":"","type":"string"}},"type":"object"}
...
    AutomationSettings: {"properties":{"enabled":{"type":"boolean"}},"required":["enabled"],"type":"object"}
...
    CoachingSettings: {"properties":{"enabled":{"type":"boolean"},"locale":{"type":"string"},"overlay_mode":{"type":"string"},"profiles":{"additionalProperties":{"$ref":"#/components/schemas/CoachingProfileSettings"},"default":{},"type":"object"},"quiet_hours":{"default":[],"items":{"$ref":"#/components/schemas/CoachingTimeRangeSettings"},"type":"array"},"regime_goals":{"additionalProperties":{"format":"uint32","minimum":0,"type":"integer"},"default":{},"type":"object"},"tone":{"type":"string"}},"required":["enabled","tone","locale","overlay_mode"],"type":"object"}
...
    DailySummaryResponse: {"properties":{"cpu_avg":{"format":"double","type":"number"},"date":{"type":"string"},"events_logged":{"format":"uint64","minimum":0,"type":"integer"},"frames_captured":{"format":"uint64","minimum":0,"type":"integer"},"memory_avg_percent":{"format":"double","type":"number"},"top_apps":{"items":{"$ref":"#/components/schemas/AppUsageEntry"},"type":"array"},"total_active_secs":{"format":"uint64","minimum":0,"type":"integer"},"total_idle_secs":{"format":"uint64","minimum":0,"type":"integer"}},"required":["date","total_active_secs","total_idle_secs","top_apps","cpu_avg","memory_avg_percent","frames_captured","events_logged"],"type":"object"}
...
    ErrorResponse: {"properties":{"code":{"type":"string"},"error":{"type":"string"},"status":{"format":"uint16","maximum":65535,"minimum":0,"type":"integer"}},"required":["code","error","status"],"type":"object"}
...
    FocusAutoSettings: {"properties":{"cooldown_secs":{"default":300,"format":"uint64","minimum":0,"type":"integer"},"duration_minutes":{"default":25,"format":"uint32","minimum":0,"type":"integer"},"enabled":{"default":false,"type":"boolean"},"trigger_apps":{"default":[],"items":{"type":"string"},"type":"array"},"trigger_schedules":{"default":[],"items":{"$ref":"#/components/schemas/FocusScheduleSettings"},"type":"array"}},"type":"object"}
...
    IndicatorSettings: {"properties":{"border_opacity":{"format":"float","type":"number"},"show_border":{"type":"boolean"},"show_panel":{"type":"boolean"}},"required":["show_border","show_panel","border_opacity"],"type":"object"}
...
    IntegrationSettings: {"properties":{"auth_profile_kind":{"type":"string"},"enabled":{"type":"boolean"},"request_timeout_secs":{"format":"uint64","minimum":0,"type":"integer"},"sync_interval_secs":{"format":"uint64","minimum":0,"type":"integer"}},"required":["enabled","auth_profile_kind","request_timeout_secs","sync_interval_secs"],"type":"object"}
...
    MonitorControlSettings: {"properties":{"input_activity":{"type":"boolean"},"privacy_mode":{"type":"boolean"},"process_monitoring":{"type":"boolean"}},"required":["process_monitoring","input_activity","privacy_mode"],"type":"object"}
    NetworkSettings: {"properties":{"grpc_enabled":{"type":"boolean"},"grpc_endpoint":{"type":"string"},"request_timeout_ms":{"format":"uint64","minimum":0,"type":"integer"},"server_base_url":{"type":"string"},"tls_enabled":{"type":"boolean"}},"required":["server_base_url","request_timeout_ms","grpc_enabled","grpc_endpoint","tls_enabled"],"type":"object"}
    NotificationSettings: {"properties":{"enabled":{"type":"boolean"},"high_usage_notification":{"type":"boolean"},"high_usage_threshold":{"format":"uint32","minimum":0,"type":"integer"},"idle_notification":{"type":"boolean"},"idle_notification_mins":{"format":"uint32","minimum":0,"type":"integer"},"long_session_mins":{"format":"uint32","minimum":0,"type":"integer"},"long_session_notification":{"type":"boolean"}},"required":["enabled","idle_notification","idle_notification_mins","long_session_notification","long_session_mins","high_usage_notification","high_usage_threshold"],"type":"object"}
...
    PrivacySettings: {"properties":{"auto_exclude_sensitive":{"type":"boolean"},"excluded_app_patterns":{"items":{"type":"string"},"type":"array"},"excluded_apps":{"items":{"type":"string"},"type":"array"},"excluded_title_patterns":{"items":{"type":"string"},"type":"array"},"pii_filter_level":{"type":"string"}},"required":["excluded_apps","excluded_app_patterns","excluded_title_patterns","auto_exclude_sensitive","pii_filter_level"],"type":"object"}
...
    SearchResponse: {"properties":{"limit":{"format":"uint","minimum":0,"type":"integer"},"offset":{"format":"uint","minimum":0,"type":"integer"},"query":{"type":"string"},"results":{"items":{"$ref":"#/components/schemas/SearchResult"},"type":"array"},"total":{"format":"uint64","minimum":0,"type":"integer"}},"required":["query","total","offset","limit","results"],"type":"object"}
    SearchResult: {"properties":{"app_name":{"type":["string","null"]},"id":{"type":"string"},"image_url":{"type":["string","null"]},"importance":{"format":"float","type":["number","null"]},"matched_text":{"type":["string","null"]},"result_type":{"type":"string"},"tags":{"items":{"$ref":"#/components/schemas/TagInfo"},"type":["array","null"]},"timestamp":{"type":"string"},"window_title":{"type":["string","null"]}},"required":["result_type","id","timestamp"],"type":"object"}
...
    SemanticSearchResult: {"description":"A single semantic search result, enriched with segment metadata.","properties":{"content_label":{"type":["string","null"]},"content_type":{"type":"string"},"dominant_category":{"type":["string","null"]},"duration_secs":{"format":"uint64","minimum":0,"type":["integer","null"]},"llm_summary":{"type":["string","null"]},"original_text":{"type":"string"},"regime_label":{"type":["string","null"]},"score":{"format":"float","type":"number"},"segment_end":{"type":["string","null"]},"segment_id":{"type":"string"},"segment_start":{"type":["string","null"]},"similarity":{"format":"float","type":"number"},"time_decay":{"format":"float","type":"number"},"timestamp":{"type":"string"}},"required":["segment_id","content_type","original_text","score","similarity","time_decay","timestamp"],"type":"object"}
...
    SuggestionSettings: {"properties":{"enabled":{"type":"boolean"}},"required":["enabled"],"type":"object"}
...
    TelemetrySettings: {"properties":{"crash_reports":{"type":"boolean"},"enabled":{"type":"boolean"},"performance_metrics":{"type":"boolean"},"usage_analytics":{"type":"boolean"}},"required":["enabled","crash_reports","usage_analytics","performance_metrics"],"type":"object"}
...
    TimelineResponse: {"properties":{"items":{"items":{"$ref":"#/components/schemas/TimelineItem"},"type":"array"},"segments":{"items":{"$ref":"#/components/schemas/AppSegment"},"type":"array"},"session":{"$ref":"#/components/schemas/SessionInfo"}},"required":["session","items","segments"],"type":"object"}
...